Move ops tests
